priority queue now working correctly, also accounting for min frame rate and limiting user app update time to this - although actual time in update function cannot be directly controlled

+++ DreamOS/DreamApp.cpp
@@ -21,12 +21,12 @@ RESULT DreamAppBase::IncrementTimeRun(double usTimeDelta) {
 	return R_PASS;
 }
 
-float DreamAppBase::GetTimeRun() {
+double DreamAppBase::GetTimeRun() {
 	return m_usTimeRun;
 }
 
 // In short, this will return 
-float DreamAppBase::GetEffectivePriorityValue() const {
+double DreamAppBase::GetEffectivePriorityValue() const {
 	return (m_usTimeRun * m_priority);
 }
